+++ cli/internal/cmd/configgenerate.go
@@ -20,7 +20,7 @@ import (
 
 func newConfigGenerateCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "generate {aws|azure|gcp}",
+		Use:   "generate {aws|azure|gcp|qemu}",
 		Short: "Generate a default configuration file",
 		Long:  "Generate a default configuration file for your selected cloud provider.",
 		Args: cobra.MatchAll(
@@ -54,6 +54,11 @@ func configGenerate(cmd *cobra.Command, fileHandler file.Handler, provider cloud
 
 	conf := config.Default()
 	conf.RemoveProviderExcept(provider)
+
+	// set a lower default for QEMU's state disk
+	if provider == cloudprovider.QEMU {
+		conf.StateDiskSizeGB = 10
+	}
 
 	if flags.file == "-" {
 		content, err := encoder.NewEncoder(conf).Encode()
@@ -90,7 +95,7 @@ func parseGenerateFlags(cmd *cobra.Command) (generateFlags, error) {
 func generateCompletion(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
 	switch len(args) {
 	case 0:
-		return []string{"aws", "gcp", "azure"}, cobra.ShellCompDirectiveNoFileComp
+		return []string{"aws", "gcp", "azure", "qemu"}, cobra.ShellCompDirectiveNoFileComp
 	default:
 		return []string{}, cobra.ShellCompDirectiveError
 	}