+++ cli/internal/terraform/terraform/qemu/outputs.tf
@@ -1,0 +1,3 @@
+output "ip" {
+  value = module.control_plane.instance_ips[0]
+}