+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Loadbalancer for control-plane recovery
 - K8s conformance mode
+- Local cluster creation based on QEMU
 
 ### Changed
 <!-- For changes in existing functionality.  -->